+++ server/src/main/java/fiit/vava/server/dao/repositories/ClientRepositorySql.java
@@ -1,0 +1,27 @@
+package fiit.vava.server.dao.repositories;
+
+import fiit.vava.server.Client;
+
+import java.util.List;
+
+public class ClientRepositorySql extends ClientRepository {
+    @Override
+    public List<Client> getNonConfirmedClients() {
+        return null;
+    }
+
+    @Override
+    public Client save(Client toSave) {
+        return null;
+    }
+
+    @Override
+    public List<Client> findAll() {
+        return null;
+    }
+
+    @Override
+    public Client findById(String id) {
+        return null;
+    }
+}